refactoring some classes and layout style
